refined file read and write on background/foreground state changes

+++ Simplified/NYPLAnnotations.swift
@@ -116,7 +116,7 @@ class NYPLAnnotations: NSObject {
         if(fileWriteSuccess) {
             NYPLAnnotations.lastReadBookQueue.cancelAllOperations()
         }
-
+        
     }
     
     func applicationDidEnterForeground() {
@@ -129,19 +129,29 @@ class NYPLAnnotations: NSObject {
         
         if let dir = NSSearchPathForDirectoriesInDomains(NSSearchPathDirectory.DocumentDirectory, NSSearchPathDomainMask.AllDomainsMask, true).first {
             let path = NSURL(fileURLWithPath: dir).URLByAppendingPathComponent(file)
-            let data: NSData? = NSData(contentsOfFile: (path?.absoluteString)!)
-            do {
-                if let jsonObject: NSDictionary = try! NSJSONSerialization.JSONObjectWithData(data!, options: NSJSONReadingOptions.MutableContainers) as? NSDictionary
-                {
-                    try! fileManager.removeItemAtURL(path!)
-                    Log.debug(#file,jsonObject.description)
-                } else {
-                    Log.error(#file,"Unable to read NYPLCirculationAnalytics.analyticsQueue data from file")
+            
+            if(fileManager.fileExistsAtPath((path?.path)!)) {
+                
+                let data: NSData? = NSData(contentsOfFile: (path?.path)!)
+                do {
+                    if let jsonObject: NSDictionary = try! NSJSONSerialization.JSONObjectWithData(data!, options: NSJSONReadingOptions.MutableContainers) as? NSDictionary
+                    {
+                        //if we have sucessfully read the file back in, remove it
+                        try! fileManager.removeItemAtURL(path!)
+                        Log.debug(#file,jsonObject.description)
+                        
+                        //TODO: load file contents back into queue
+                        
+                        
+                    } else {
+                        Log.error(#file,"Unable to read NYPLCirculationAnalytics.analyticsQueue data from file")
+                    }
                 }
+                
             }
             
         }
-
+        
         NYPLAnnotations.lastReadBookQueue.suspended = !NYPLAnnotations.isReachable
     }
     
@@ -173,7 +183,7 @@ class NYPLAnnotations: NSObject {
         
         lastReadBookQueue.addOperation(lastReadBookOperation)
     }
-
+    
     class func syncLastRead(book:NYPLBook, completionHandler: (responseObject: String?,
         error: NSError?) -> ()) {
